pyproject.toml
[project]
name = "django-forum"
version = "0.1.0"
description = ""
authors = [
    { name = "Kirill Tarasov", email = "bravekirty@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
# PROD
dependencies = [
    "aiohttp>=3.11.16,<4.0.0",
    "django>=5.2,<6.0",
    "pillow>=11.1.0,<12.0.0",
    "python-decouple>=3.8,<4.0",
    "requests>=2.32.3,<3.0.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "gunicorn>=23.0.0,<24.0.0",
]

[project.urls]
Homepage = "https://pet-forum-temp.onrender.com/"
Documentation = "https://github.com/Tarasyonok/pet-forum-temp"
Repository = "https://github.com/Tarasyonok/pet-forum-temp"

[tool.django]
settings_module = "django_forum.settings"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# DEV
[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "~5.1.0"
ruff = "~0.14.0"
faker = "~37.11.0"

# SECURITY
[tool.poetry.group.security.dependencies]
bandit = "~1.8.3"

# RUFF CONFIGURATION (replaces both Black and isort)
[tool.ruff]
target-version = "py313"
line-length = 120

# Linting rules
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
    "UP", # pyupgrade
    "YTT", # flake8-2020
    "FBT", # flake8-boolean-trap
    "Q", # flake8-quotes
    "RUF", # ruff-specific rules
]

lint.ignore = [
    "E501", # line-too-long - handled by formatter
    "RUF012", # mutable-class-default
]

# File exclusions
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # allow unused imports in __init__.py
"migrations/*.py" = ["F401"]  # allow unused imports in migrations